Fix #2: delaying generation of conversion expressions
The conversion expressions differ depending whether they refer to
converters defined in the same module or not, and the module where
they appear can change when one module has to be closed because it is
large enough and another module has to be started. Therefore, the
generation of the reference to the converters has to be delayed to the
point where we are sure about in which module the expression will
appear.

Reported by @nilsbecker.
https://github.com/thierry-martinez/ocaml-in-python/issues/2